[TUBEMQ-262]Create C++ flow control handler

+++ tubemq-client-twins/tubemq-client-cpp/inc/const_config.h
@@ -66,7 +66,7 @@ static const int kRebWaitPeriodWhenShutdownMs = 10000;
 // max int value
 static const int kMaxIntValue = 0x7fffffff;
 // max long value
-static const int kMaxLongValue = 0x7fffffffffffffffL;
+static const long kMaxLongValue = 0x7fffffffffffffffL;
 
 // invalid value
 static const int kInvalidValue = -2;
